+++ android/src/test/kotlin/com/posthog/posthog_flutter/PosthogFlutterPluginTest.kt
@@ -14,16 +14,6 @@ import org.mockito.Mockito
  */
 
 internal class PosthogFlutterPluginTest {
-    @Test
-    fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
-        val plugin = PosthogFlutterPlugin()
-
-        val call = MethodCall("getPlatformVersion", null)
-        val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
-        plugin.onMethodCall(call, mockResult)
-
-        Mockito.verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE)
-    }
 
     @Test
     fun onMethodCall_identify_returnsExpectedValue() {