+++ CHANGELOG.md
@@ -74,4 +74,4 @@
 
 ## 1.9.0
 
-- Posthog client library for Flutter is released!
+- PostHog client library for Flutter is released!